requirements.txt
Flask==3.0.2
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
requests-cache==1.1.1
retry-requests==2.0.0
psutil==5.9.8
Werkzeug==3.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
click==8.1.7
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.1
six==1.16.0
python-dateutil==2.8.2
pytz==2024.1
numpy==2.2.1
pandas==2.2.1

# Testing dependencies
pytest==8.0.2
pytest-flask==1.3.0
pytest-mock==3.12.0

# Server dependencies
openmeteo-requests==1.3.0
openmeteo-sdk>=1.1.0
supervisor==4.2.5

# Development tools
black==24.2.0
isort==5.13.2
flake8==7.0.0